Add files via upload
The Lenet-5 network consists of three Convolution layers, two Maxpooling, FullyConnected and three Relu, and in this project, the functions nn.conv2d, nn.linear, nn.relu, and nn.maxpool2d are used, and Train and Test are The network is implemented using Dataset Mnist. And after Train the network, the weights are saved with the .pt extension in Google Collab or on the server. At the end of the Test section, Accuracy (99.1) of the network is printed.